requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
httpx==0.26.0
oss2==2.18.3
python-dotenv==1.0.0
pydantic==2.5.2
typing-extensions==4.8.0
aiofiles==23.2.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1 